requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.5
aiosignal==1.4.0
altair==5.5.0
annotated-types==0.7.0
anyio==4.13.0
asyncio==4.0.0
attrs==26.1.0
blinker==1.9.0
cachetools==5.5.2
certifi==2026.2.25
charset-normalizer==3.4.7
click==8.3.2
colorama==0.4.6
contourpy==1.3.3
cycler==0.12.1
disutils==1.4.32.post2
fastapi==0.108.0
fonttools==4.62.1
frozenlist==1.8.0
gitdb==4.0.12
GitPython==3.1.46
h11==0.16.0
idna==3.11
imbalanced-learn==0.14.1
importlib_metadata==7.2.1
Jinja2==3.1.6
joblib==1.3.2
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kiwisolver==1.5.0
lightgbm==4.6.0
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.8
mdurl==0.1.2
multidict==6.7.1
narwhals==2.19.0
numpy==1.26.4
packaging==23.2
pandas==2.1.4
pillow==10.4.0
propcache==0.4.1
protobuf==4.25.9
py4j==0.10.9.7
pyarrow==23.0.1
pydantic==2.5.3
pydantic_core==2.14.6
pydeck==0.9.2
Pygments==2.20.0
pyparsing==3.3.2
pyspark==3.5.0
python-dateutil==2.9.0.post0
pytz==2026.1.post1
referencing==0.37.0
requests==2.31.0
rich==13.9.4
rpds-py==0.30.0
scikit-learn==1.8.0
scipy==1.17.1
seaborn==0.13.2
setuptools==69.5.1
six==1.17.0
sklearn-compat==0.1.5
smmap==5.0.3
starlette==0.32.0.post1
streamlit==1.30.0
tenacity==8.5.0
threadpoolctl==3.6.0
toml==0.10.2
tornado==6.5.5
typing_extensions==4.15.0
tzdata==2026.1
tzlocal==5.3.1
urllib3==2.6.3
uvicorn==0.25.0
validators==0.35.0
watchdog==6.0.0
yarl==1.23.0
zipp==3.23.1

kafka-python
pyarrow
mlflow
